requirements.txt
graphviz==0.20.1
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
rich==13.5.3
typer==0.9.0
typing_extensions==4.8.0
